Add tech for generating medals from the base medal logo
Uses CSS blending to apply colors to the medals. Kind of fragile since it can be affected by the background color of whatever element it is being used in, but it's neat and I like it so I'm keeping it for now.

Might create some static pre-renders of the bronze, silver, gold, and diamond medals if it becomes too much of a hassle.